fix: Lottos 객체 제거
