+++ src/main/kotlin/com/nextstep/lotto/LottoApplication.kt
@@ -1,14 +1,14 @@
 package com.nextstep.lotto
 
+import com.nextstep.lotto.domain.LottoFactory
 import com.nextstep.lotto.domain.LottoResult
-import com.nextstep.lotto.domain.Lottos
 import com.nextstep.lotto.domain.WinningLotto
 import com.nextstep.lotto.view.LottoInputView
 import com.nextstep.lotto.view.LottoOutputView
 
 fun main() {
     val price = LottoInputView.inputPrice()
-    val lottos = Lottos.buyLotto(price)
+    val lottos = LottoFactory.buyLotto(price)
     LottoOutputView.showBuyResult(lottos)
 
     val winningNumbers = LottoInputView.inputWinningNumbers()
